+++ tests/drawing/cairo/utils.py
@@ -39,7 +39,7 @@ def find_open_image_png_function():
     except ImportError:
         pass
 
-    raise ImportError('PIL or OpenCV required to run Cairo tests')
+    raise ImportError('PIL+NumPy or OpenCV required to run Cairo tests')
 
 
 def find_image_comparison():
@@ -49,6 +49,18 @@ def find_image_comparison():
     if cairo is None:
         return dummy_comparison
     return image_comparison
+
+
+def are_tests_supported():
+    if cairo is None:
+        return False, "cairo not found"
+
+    try:
+        find_open_image_png_function()
+    except ImportError:
+        return False, "PIL+NumPy or OpenCV not found"
+
+    return True, ""
 
 
 def _load_image(filename, fmt):