+++ tests/drawing/matplotlib/test_graph.py
@@ -1,3 +1,4 @@
+import os
 import random
 import unittest
 
@@ -117,6 +118,16 @@ class GraphTestRunner(unittest.TestCase):
         g = Graph.Ring(5)
         with overridden_configuration("plotting.backend", "matplotlib"):
             plot(g, target="graph_basic.png", layout=self.layout_small_ring)
+            os.unlink("graph_basic.png")
+
+    @image_comparison(baseline_images=["graph_with_curved_edges"])
+    def test_graph_with_curved_edges(self):
+        plt.close("all")
+        g = Graph.Ring(24, directed=True, mutual=True)
+        fig, ax = plt.subplots()
+        lo = g.layout("circle")
+        lo.scale(3)
+        plot(g, target=ax, layout=lo)
 
 
 class ClusteringTestRunner(unittest.TestCase):